+++ src/hooks/useNav.ts
@@ -1,5 +1,5 @@
 import {useNavigation} from '@react-navigation/native';
-import {RootStackParamList} from '../stacks/RootNavigation';
+import {RootStackParamList} from '../navigations/RootNavigation';
 import {StackNavigationProp} from '@react-navigation/stack';
 
 export const useNav = <T extends keyof RootStackParamList>() =>
