+++ src/navigations/RootNavigation.tsx
@@ -2,9 +2,11 @@ import React from 'react';
 import {createStackNavigator} from '@react-navigation/stack';
 import MainTab from '../components/MainTabBar';
 import UserInfoEditScreen from '../screens/UserInfoEditScreen';
+import CageTopTab from './CageTopTab';
 
 export type RootStackParamList = {
   MainTab: undefined;
+  CageTopTab: undefined;
   UserInfoEditScreen: {
     avatarUrl: string;
     nickname: string;
@@ -15,7 +17,7 @@ export type RootStackParamList = {
 
 const Stack = createStackNavigator<RootStackParamList>();
 
-const inputScreenHeaderOption = {
+export const inputScreenHeaderOption = {
   headerTitle: ' ',
   headerShadowVisible: false,
 };
@@ -29,6 +31,13 @@ const RootNavigation = () => {
         options={{
           ...inputScreenHeaderOption,
           headerShown: false,
+        }}
+      />
+      <Stack.Screen
+        name={'CageTopTab'}
+        component={CageTopTab}
+        options={{
+          ...inputScreenHeaderOption,
         }}
       />
       <Stack.Screen
